Added tests whenever there is a config file or not
